KVM: x86/pmu: Allow accessing global registers directly in passthrough mode
There can be cases where fewer than actually available hardware counters
are advertised to the guest. In such cases, guest accesses to the global
control and status registers are intercepted in order to disallow the
guest from programming hidden bits. When the PMU context is switched at
the vcpu boundary, the state of the global registers stashed in struct
kvm_pmu may be out of date. Always maintain the latest state through the
following means.
  * Read hardware global status for RDMSR interceptions
  * Write hardware global control for WRMSR interceptions
  * Write hardware global status clear for WRMSR interceptions

Signed-off-by: Sandipan Das <sandipan.das@amd.com>