perf/x86/amd/core: Fix NMI latency workaround for legacy counters
Due to NMI latency, an overflow interrupt can arrive after a counter has
been disabled. In such cases, the interrupt handler cannot attribute the
overflow to any known active counter and the kernel assumes it to be an
unknown NMI. Commit df4d29732fda ("perf/x86/amd: Change/fix NMI latency
mitigation to use a timestamp") fixes the issue but only for  processors
which have X86_FEATURE_PERFCTR_CORE i.e. have bit 23 (PerfCtrExtCore) of
CPUID leaf 0x80000001 ECX set.

In general, not all processors which have this feature suffer from NMI
latency but then, a KVM guest can be started on affected hardware in a
way that the PerfCtrExtCore CPUID bit is not advertised to the guest. In
this case, the guest kernel will not use the NMI latency mitigation and
receive unknown NMIs. Use the mitigation irrespective of the presence of
the PerfCtrExtCore bit.

Fixes: df4d29732fda ("perf/x86/amd: Change/fix NMI latency mitigation to use a timestamp")
Signed-off-by: Sandipan Das <sandipan.das@amd.com>